pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["NefariousNiru <nirupomboseroy99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
langchain = "^0.3.25"
tiktoken = "^0.9.0"
openai = "^1.88.0"
pydantic = "^2.11.7"
python-dotenv = "^1.1.0"
pinecone = "^7.2.0"
python-multipart = "^0.0.20"
langchain-community = "^0.3.25"
grpcio = "^1.73.0"
grpcio-tools = "^1.73.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
